Update and rename configure.sh to docker-entrypoint.sh

+++ docker-entrypoint.sh
@@ -13,10 +13,16 @@ if [[ "$SAMBA_DOMAIN" != "tld" && "$SAMBA_DNS_REALM" != "tld.your.domain" ]]; th
         --realm="$SAMBA_DNS_REALM" \
         --dns-backend="SAMBA_INTERNAL"
     echo "$SAMBA_DOMAIN - Domain $SAMBA_DC_ACT Successfully."
-    
+  fi
+  if [ ! -f /etc/krb5.conf ]; then
     ln -sf /var/lib/samba/private/krb5.conf /etc/krb5.conf
   fi      
+  if [ -f /etc/samba/smb.conf ]; then
+    exec /usr/sbin/samba -i
+  fi
+  if [ -f /etc/ntpd.conf ]; then
+    exec ntpd -d
+  fi
 fi
 
-exec /usr/sbin/samba -i
-exec ntpd -d
+exec "$@"
